Moved unit test to appropriate location and added import guards.
